requirements.txt
requests>=2.31.0
isodate>=0.6.1
python-dotenv>=1.0.0
pyyaml>=6.0.1
# テスト用
pytest>=7.4.0
flake8>=6.0.0
